setting radius to double

+++ api/Entities/Location.cs
@@ -12,7 +12,7 @@ namespace api.Entities
         public string UserName { get; set; } = string.Empty;
         public string Title { get; set; } = string.Empty;
         public string StreetAddress { get; set; } = string.Empty;
-        public int Radius { get; set; }
+        public double Radius { get; set; }
         public string RadiusUnit { get; set; } = string.Empty;
         public List<Reminder> Reminders { get; set; } = new List<Reminder>();
